- Upgrade Kotlin version to the latest. - Cleanup the OffsetBasedPageRequest class. - Add .editorconfig. - Use let more.

+++ src/main/kotlin/org/motivepick/service/TaskService.kt
@@ -16,7 +16,7 @@ interface TaskService {
 
     fun softDeleteTaskById(taskId: Long): TaskView?
 
-    fun findForCurrentUser(listType: TaskListType, offset: Int, limit: Int): Page<TaskView>
+    fun findForCurrentUser(listType: TaskListType, offset: Long, limit: Int): Page<TaskView>
 
     fun findScheduleForCurrentUser(): ScheduleView
